+++ sdk/ai/ai-projects/src/api/index.ts
@@ -6,7 +6,7 @@ export { createAIProject, AIProjectContext } from "./aiProjectContext.js";
 export { AIProjectClientOptionalParams };
 
 /** Options for uploading datasets */
-export interface DatasetUploadOptionalOptions {
+export interface DatasetUploadOptions {
   /**  The name of an Azure Storage Account connection, where the file should be uploaded. If not specified, the default Azure Storage Account connection will be used.  */
   connectionName?: string;
   /** Only for folder uploads. A regex pattern to filter files to be uploaded. Only files matching the pattern will be uploaded. If not specified, all files will be uploaded. */
@@ -17,6 +17,6 @@ export interface DatasetUploadOptionalOptions {
  * @internal
  * Options for dataset upload operations that are internal and may not be exposed to all users
  */
-export interface DatasetUploadInternalOptionalParams extends DatasetUploadOptionalOptions {
+export interface DatasetUploadInternalOptions extends DatasetUploadOptions {
   projectOptions?: AIProjectClientOptionalParams;
 }